# Src/alex-pi/src/app/_api/models/web-event-log.ts
/* tslint:disable */
import { WebsiteUser } from './website-user';
export interface WebEventLog {
  id?: number;
  websiteUserId?: number;
  eventName?: string;
  doneAt?: string;
  websiteUser?: WebsiteUser;
  browserSignature?: string;
}