# Src/alex-pi/swagger.v5.0.0-rc2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "description": "Demo API for swagger code generation",
    "version": "v1"
  },
  "paths": {
    "/api/Ocr/{imgurl}": {
      "get": {
        "tags": ["Ocr"],
        "operationId": "Get",
        "parameters": [
          {
            "name": "imgurl",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": ["WeatherForecast"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/WebEventLogs": {
      "get": {
        "tags": ["WebEventLogs"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebEventLog"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebEventLog"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebEventLog"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["WebEventLogs"],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebEventLogs/{id}": {
      "get": {
        "tags": ["WebEventLogs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["WebEventLogs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WebEventLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["WebEventLogs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebEventLog"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebsiteUsers": {
      "get": {
        "tags": ["WebsiteUsers"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebsiteUser"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebsiteUser"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebsiteUser"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["WebsiteUsers"],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebsiteUsers/{id}": {
      "get": {
        "tags": ["WebsiteUsers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["WebsiteUsers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["WebsiteUsers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteUser"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "WebsiteUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "eventData": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "visitCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "note": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastVisitAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "webEventLog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebEventLog"
            }
          },
          "visitCount__New": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WebEventLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "websiteUserId": {
            "type": "integer",
            "format": "int32"
          },
          "eventName": {
            "type": "string"
          },
          "doneAt": {
            "type": "string",
            "format": "date-time"
          },
          "nickname": {
            "type": "string"
          },
          "websiteUser": {
            "$ref": "#/components/schemas/WebsiteUser"
          },
          "browserSignature": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}